+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added locations feature
 - Added Phone number parse feature
 - Removed depreciated class and method
+- Added operator overloading on nepalidate
 
 ## v0.5.6 - (July 7, 2022)
 - Test for automatic package publish